[UI] Fix PokeCenter Roster Tab UI Inconsistencies
The UI would mess up upon altering your roster in any way. This is, in part, due to how poorly the PokeCenter is coded, but nevertheless, this will fix the issue for now. Ideally, I'm thinking that these requests should be handled in JSON, with the appropriate divs updated accordingly.